96 general fixes 1 (#97)
* Adjutsed required fields

* Adjusted README

* Added info section to profile

* Added info bank navpanel button

* Pretty up user events page

* Added membership expiration check before buying ticket

* Extended membership by membership duration when renewing

* Adjusted accounts tab text

* Fixed task edit bug

* Added required field comps

* Implemented start and end time utils

* Adjusted delete permission

* Implemented util for dummy id

* Enabled autifilling task shift

* Enabled deleting all task shifts for a task

* Added tooltip

* Fixed sorting bug

* Sorted tasks in kanban board

* Enabled creating tasks and shifts in kanban board

* Added unassigned option to task filter

* Remove console.logs

+++ src/app/lib/user-actions.ts
@@ -9,6 +9,25 @@ import { decryptJWT, encryptJWT, generateUserCredentials, getUserByUniqueKey } f
 import { sendUserCredentials } from "./mail-service/mail-service";
 import { LoginSchema, ResetCredentialsSchema } from "./definitions";
 
+export const getUserById = async (
+    currentState: DatagridActionState,
+    userId: string,
+): Promise<DatagridActionState> => {
+    const newActionState: DatagridActionState = { ...currentState };
+    try {
+        const user = await prisma.user.findUniqueOrThrow({
+            where: { id: userId },
+        });
+        newActionState.status = 200;
+        newActionState.result = [user];
+    } catch (error) {
+        newActionState.status = 500;
+        newActionState.errorMsg = error.message;
+        newActionState.result = [];
+    }
+    return newActionState;
+};
+
 export const createUser = async (
     currentActionState: FormActionState,
     fieldValues: Prisma.UserCreateInput,
@@ -45,9 +64,11 @@ export const getAllUsers = async (
         });
         newActionState.status = 200;
         newActionState.result = users;
+        newActionState.errorMsg = "";
     } catch (error) {
         newActionState.status = 500;
         newActionState.errorMsg = error.message;
+        newActionState.result = [];
     }
     return newActionState;
 };
@@ -237,6 +258,19 @@ export const deleteUser = async (
 ): Promise<FormActionState> => {
     const newActionState = { ...currentActionState };
     try {
+        const adminCount = await prisma.user.count({
+            where: {
+                role: GlobalConstants.ADMIN,
+            },
+        });
+
+        if (adminCount <= 1) {
+            newActionState.status = 400;
+            newActionState.errorMsg =
+                "You are the last admin standing. Find an heir before leaving.";
+            newActionState.result = "";
+            return newActionState;
+        }
         const deleteCredentials = prisma.userCredentials.deleteMany({
             where: {
                 email: user.email as string,
@@ -265,7 +299,7 @@ export const deleteUser = async (
     return newActionState;
 };
 
-export const getUserParticipantEvents = async (
+export const getUserEvents = async (
     userId: string,
     currentState: DatagridActionState,
 ): Promise<DatagridActionState> => {
@@ -273,11 +307,18 @@ export const getUserParticipantEvents = async (
     try {
         const events = await prisma.event.findMany({
             where: {
-                participantUsers: {
-                    some: {
-                        userId: userId,
+                OR: [
+                    {
+                        participantUsers: {
+                            some: {
+                                userId: userId,
+                            },
+                        },
                     },
-                },
+                    {
+                        hostId: userId,
+                    },
+                ],
             },
         });
         newActionState.status = 200;
